+++ tests/unit/test_transaction.py
@@ -55,6 +55,16 @@ class TestTransaction(unittest.TestCase):
         self.assertEqual(transaction.disbursement_details.funds_held, False)
         self.assertEqual(transaction.is_disbursed, True)
 
+    def test_constructor_includes_sepa_direct_debit_return_code(self):
+        attributes = {
+            'amount': '27.00',
+            'sepa_direct_debit_return_code': 'AM04'
+        }
+
+        transaction = Transaction(None, attributes)
+
+        self.assertEqual(transaction.sepa_direct_debit_return_code, 'AM04')
+
     def test_transaction_handles_nil_risk_data(self):
         attributes = {
             'amount': '27.00',
@@ -217,3 +227,12 @@ class TestTransaction(unittest.TestCase):
 
         transaction = Transaction(None, attributes)
         self.assertEqual(transaction.installments["count"], 4)
+
+    def test_gateway_rejection_reason_for_excessive_retry(self):
+        attributes = {
+            'amount': '27.00',
+            'gateway_rejection_reason': 'excessive_retry'
+        }
+
+        transaction = Transaction(None, attributes)
+        self.assertEqual(transaction.gateway_rejection_reason, braintree.Transaction.GatewayRejectionReason.ExcessiveRetry)